Moved registry methods to thier own class
