+++ src/main/java/notreepunching/proxy/ClientProxy.java
@@ -9,7 +9,9 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import net.minecraftforge.client.event.ModelRegistryEvent;
 import net.minecraftforge.client.model.ModelLoader;
+import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.relauncher.Side;
 import notreepunching.NoTreePunching;
@@ -17,6 +19,7 @@ import notreepunching.block.ModBlocks;
 import notreepunching.item.ModItems;
 import notreepunching.client.particle.FirepitParticle;
 import notreepunching.client.particle.FirepitSmokeParticle;
+import notreepunching.registry.ClientRegistryHandler;
 
 import java.util.Random;
 
@@ -24,10 +27,9 @@ import java.util.Random;
 @Mod.EventBusSubscriber(Side.CLIENT)
 public class ClientProxy extends CommonProxy {
 
-    @SubscribeEvent
-    public static void registerModels(ModelRegistryEvent event) {
-        ModItems.registerItemModels();
-        ModBlocks.registerItemBlockModels();
+    @Override
+    public void preInit(FMLPreInitializationEvent event){
+        MinecraftForge.EVENT_BUS.register(new ClientRegistryHandler());
     }
 
     @Override
