+++ src/main/java/notreepunching/proxy/ServerProxy.java
@@ -1,8 +1,13 @@
 package notreepunching.proxy;
 
+import net.minecraft.util.text.translation.I18n;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.relauncher.Side;
 
 @Mod.EventBusSubscriber(Side.SERVER)
 public class ServerProxy extends CommonProxy {
+
+    public String localize(String unlocalized, Object... args) {
+        return I18n.translateToLocalFormatted(unlocalized, args);
+    }
 }
